refactored library, fixed GH CI, added type annotations
