+++ .github/workflows/pytest.yml
@@ -5,19 +5,20 @@ on: [push]
 jobs:
   build:
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.8", "3.9", "3.10"]
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v3
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Setup Redis
+      run: |
+        sudo apt install redis
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install pytest
     - name: Analysing the code with pylint
       run: |
+        nohup redis  > /dev/null &
         pytest -v
